requirements.txt
openai-whisper
torch
openai
tqdm
